refactor(species-id): extract geo index into cell-csr-index crate
Move src/geo_index.rs into a new, domain-agnostic crate: a zero-copy,
mmap-backed index mapping H3 cells to sorted u32 ID lists with O(log n)
point lookup. Renamed species->id terminology (GeoIndex -> CellCsrIndex,
species_at -> ids_at), gave it a self-contained CellIndexError, and made
the stale-count check optional (load takes Option<usize>).

observing-species-id depends on the new crate and drops its now-unused
h3o/memmap2 deps. The OGI1 binary format is unchanged, so the existing
build pipeline artifacts load as-is.
