+++ crates/cell-csr-index/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "cell-csr-index"
+version = "0.1.0"
+edition = "2021"
+description = "Zero-copy, mmap-backed index mapping H3 cells to sorted u32 ID lists, with O(log n) point lookup"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+# H3 geospatial cells
+h3o = { version = "0.9", default-features = false }
+
+# Zero-copy mmap + slice reinterpretation
+memmap2 = "0.9"
+bytemuck = { version = "1", features = ["derive"] }
+
+# Logging
+tracing = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
