+++ crates/observing-species-id/src/main.rs
@@ -4,7 +4,6 @@
 
 mod embeddings;
 mod error;
-mod geo_index;
 mod model;
 mod preprocessing;
 mod server;
